Added translator syntax for state machines.
State machines can now be defined in a .pysm file.  As long as the
StateMachines module has been imported before importing a state machine
defined in a .pysm file, the State Machines preprocessor will be able to
transparently translate that file into a proper pure-python module.

Pysm files introduce two new keywords to the Python syntax: "State" and
"Translate".  They can be thought of as aliases for "class" and "def"
plus the appropriate overhead for defining a state.  See the pysmTest.pysm
file for an example.

git-svn-id: https://www.lri.fr/svn/in-situ/code/PythonStateMachines/trunk@899 2173f847-0edd-4382-bdcc-49b71dc1bf57